refactor: move the generic request handler to separate file
