+++ src/express.ts
@@ -9,33 +9,17 @@ import {
   getScenarios as apiGetScenarios,
 } from './apis';
 import {
-  getGraphQlMocks,
-  getGraphQlMock,
-  createGraphQlRequestHandler,
-} from './graph-ql';
-import {
-  getHttpMocks,
-  getHttpMockAndParams,
-  createHttpRequestHandler,
-} from './http';
-import {
-  Mock,
   Options,
   ScenarioMap,
   DefaultScenario,
   Context,
   Scenario,
-  PartialContext,
   InternalRequest,
-  Result,
 } from './types';
 import { getUi, updateUi } from './ui';
-import {
-  getScenarioIdsFromCookie,
-  getDataMocksServerCookie,
-  setDataMocksServerCookie,
-} from './cookies';
+import { getScenarioIdsFromCookie, setDataMocksServerCookie } from './cookies';
 import { getContextFromScenarios } from './utils/get-context-from-scenarios';
+import { handleRequest } from './handle-request';
 
 export { createExpressApp };
 
@@ -210,35 +194,6 @@ function createExpressApp({
   }
 }
 
-function updateContext(context: Context, partialContext: PartialContext) {
-  const newContext: Context = {
-    ...context,
-    ...(typeof partialContext === 'function'
-      ? partialContext(context)
-      : partialContext),
-  };
-
-  return newContext;
-}
-
-function mergeMocks(scenarioMap: ({ mocks: Mock[] } | Mock[])[]) {
-  return scenarioMap.reduce<Mock[]>(
-    (result, scenarioMock) =>
-      result.concat(
-        Array.isArray(scenarioMock) ? scenarioMock : scenarioMock.mocks,
-      ),
-    [],
-  );
-}
-
-function getMocksFromScenarios(scenarios: Scenario[]) {
-  const mocks = mergeMocks(scenarios);
-  const httpMocks = getHttpMocks(mocks);
-  const graphQlMocks = getGraphQlMocks(mocks);
-
-  return { httpMocks, graphQlMocks };
-}
-
 function getScenarios({
   defaultScenario,
   scenarioMap,
@@ -251,98 +206,6 @@ function getScenarios({
   return [defaultScenario].concat(
     scenarioIds.map(scenarioId => scenarioMap[scenarioId]),
   );
-}
-
-async function handleRequest({
-  req,
-  getServerSelectedScenarioIds,
-  defaultScenario,
-  scenarioMap,
-  getServerContext,
-  setServerContext,
-  getCookie,
-  cookieMode,
-  setCookie,
-}: {
-  req: InternalRequest;
-  getServerSelectedScenarioIds: () => string[];
-  defaultScenario: DefaultScenario;
-  scenarioMap: ScenarioMap;
-  getServerContext: () => Context;
-  setServerContext: (context: Context) => void;
-  getCookie: (cookieName: string) => string;
-  setCookie: (cookieName: string, cookieValue: string) => void;
-  cookieMode: boolean;
-}) {
-  const dataMocksServerCookie = getDataMocksServerCookie({
-    getCookie,
-    defaultScenario,
-  });
-
-  const getSelectedScenarioIds = cookieMode
-    ? () => dataMocksServerCookie.scenarios
-    : getServerSelectedScenarioIds;
-
-  const getContext = cookieMode
-    ? () => dataMocksServerCookie.context
-    : getServerContext;
-
-  const setContext = cookieMode
-    ? (context: Context) => {
-        dataMocksServerCookie.context = context;
-      }
-    : setServerContext;
-
-  const selectedScenarioIds = getSelectedScenarioIds();
-
-  const selectedScenarios = getScenarios({
-    defaultScenario,
-    scenarioMap,
-    scenarioIds: selectedScenarioIds,
-  });
-
-  const { httpMocks, graphQlMocks } = getMocksFromScenarios(selectedScenarios);
-
-  const graphQlMock = getGraphQlMock(req.path, graphQlMocks);
-
-  // Default when nothing matches
-  let result: Result = { status: 404 };
-
-  if (graphQlMock) {
-    const requestHandler = createGraphQlRequestHandler({
-      graphQlMock,
-      updateContext: localUpdateContext,
-      getContext,
-    });
-
-    result = await requestHandler(req);
-  } else {
-    const { httpMock, params } = getHttpMockAndParams(req, httpMocks);
-    if (httpMock) {
-      const requestHandler = createHttpRequestHandler({
-        httpMock,
-        params,
-        getContext,
-        updateContext: localUpdateContext,
-      });
-
-      result = await requestHandler(req);
-    }
-  }
-
-  if (cookieMode) {
-    setDataMocksServerCookie({ setCookie, value: dataMocksServerCookie });
-  }
-
-  return result;
-
-  function localUpdateContext(partialContext: PartialContext) {
-    const newContext = updateContext(getContext(), partialContext);
-
-    setContext(newContext);
-
-    return newContext;
-  }
 }
 
 function expressSetCookie(res: Response) {
